Create "chunks" module, update docs
